Multiple Indicator Cluster Survey (MICS) 2014 Household Listing questionnaire data for Punjab, Pakistan

MICS2014HL data of Household Listing, MICS Punjab, 2014.

data(MICS2014HL)

Format

A data frame with 246501 rows and 81 variables:

HH1

Cluster number

HH2

Household number

HL1

Line number

HL3

Relationship to the head

HL4

Sex

HL5M

Month of birth

HL5Y

Year of birth

HL6

Age

HL6B

Marital status of member

HL7

Line number of woman age 15 - 49

HL7B

Line number for children age 0-4

HL10A

Read in any langauge with understanding

HL10BA

Can read in any language with understanding - Urdu

HL10BB

Can read in any language with understanding - English

HL10BC

Can read in any language with understanding - Punjabi

HL10BD

Can read in any language with understanding - Saraiki

HL10BX

Can read in any language with understanding - Other (specify)

HL10BZ

Can read in any language with understanding - DK

HL10C

Write in any langauge with understanding

HL10DA

Can write in any language with understanding - Urdu

HL10DB

Can write in any language with understanding - English

HL10DC

Can write in any language with understanding - Punjabi

HL10DD

Can write in any language with understanding - Saraiki

HL10DX

Can write in any language with understanding - Other (specify)

HL10DZ

Can write in any language with understanding - DK

HL11

Is natural mother alive

HL12

Natural mother's line number in HH

HL12A

Where does natural mother live

HL13

Is natural father alive

HL14

Natural father's line number in HH

HL14A

Where does natural father live

HL15

Line number of mother or primary caretaker for children 0-14 years of age

HL16A

Had cough and fever for last 3 weeks

HL16B

Diagnosed as having TB in the past year

HL16C

Diagnosed as having hepatitis in the past year

ED1

Line number

ED3

Ever attended school or pre-school

ED4A

Highest level of education attended

ED4B

Highest grade completed at that level

ED5

Attended school during current school year (2014-2015)

ED6A

Level of education attended current school year

ED6B

Grade of education attended current school year

ED6C

Type of school attended current year (2014-15)

ED7

Attended school previous school year (2013-2014)

ED8A

Level of education attended previous school year

ED8B

Grade of education attended previous school year

ED8C

Type of school attended previous year (2013-14)

IE1

Line number

IE3

Major type of income source

IE4N

Income on daily or monthly or yearly basis: Amount

IE4U

Income on daily or monthly or yearly basis: Unit

IE5

On average, number of months worked for pay in a year

IE6

On average, number of days worked for pay in a month

IE7

Other source of income

IE8N

Additional Income on daily or monthly or yearly basis: Amount

IE8U

Additinal Income on daily or monthly or yearly basis: Unit

IE9

On average, number of months worked for additional pay in a year

IE10

On average, number of days worked for additional pay in a month

HH5D

Day of interview

HH5M

Month of interview

HH5Y

Year of interview

HH6

Area (Original)

HH7

District (Original)

MLINE

Mother's line number

FLINE

Father's line number

division

Division

hh7r

District

hh6r

Area

suburban

stratum

Straum (Original)

helevel

Education of household head

melevel

Mother's education

felevel

Father's education

schage

Age at beginning of school year

wscore

Combined wealth score

windex5

Wealth index quintile

wscoreu

Urban wealth score

windex5u

Urban wealth index quintile

wscorer

Rural wealth score

windex5r

Rural wealth index quintile

hhweight

Household sample weight

Examples

data(MICS2014HL)